+++ write_locations.py
@@ -4,39 +4,39 @@ Write one or more sets of publication location information to Thoth.
 Input: path to file containing location information, one location per line,
 containing publication ID, location platform, landing page and full text URL,
 separated by spaces.
-Requires: Thoth credentials as THOTH_EMAIL and THOTH_PWD env vars.
+Requires: Thoth personal access token as THOTH_PAT env var.
 """
 
 # Third-party package already included in thoth-dissemination/requirements.txt
-from thothlibrary import ThothClient, ThothError
+from thothlibrary import ThothError
 from os import environ
 import sys
+from thothapi import get_thoth_client
 
 
 def write_thoth_location(publication_id, location_platform, landing_page,
-                         full_text_url):
-    thoth = ThothClient()
+                         full_text_url, checksum, checksum_algorithm):
+    thoth = get_thoth_client()
     try:
-        username = environ['THOTH_EMAIL']
+        token = environ['THOTH_PAT']
     except KeyError as e:
-        raise KeyError('No Thoth username provided (THOTH_EMAIL environment variable not set)') from e
-    try:
-        password = environ['THOTH_PWD']
-    except KeyError as e:
-        raise KeyError('No Thoth password provided (THOTH_PWD environment variable not set)') from e
-    try:
-        thoth.login(username, password)
-    except ThothError:
-        raise ValueError('Thoth login failed: credentials may be incorrect')
+        raise KeyError('No Thoth token provided (THOTH_PAT environment variable not set)') from e
+
+    thoth.set_token(token)
 
     location = {
         'publicationId': publication_id,
         'landingPage': landing_page,
         'fullTextUrl': full_text_url,
         'locationPlatform': location_platform,
-        'canonical': 'false'
+        'canonical': 'false',
+        'checksum': checksum,
+        'checksumAlgorithm': checksum_algorithm
     }
-    location_id = thoth.create_location(location)
+    try:
+        location_id = thoth.create_location(location)
+    except ThothError as e:
+        raise ValueError('Failed to create location in Thoth: token may be incorrect') from e
     print(location_id)
 
 
@@ -49,6 +49,11 @@ if __name__ == '__main__':
                 # Handle locations which only have a landing page, no full text URL
                 if parts[3] == "None":
                     parts[3] = None
-                write_thoth_location(parts[0], parts[1], parts[2], parts[3])
+                # Handle locations which don't have a checksum
+                if parts[4] == "None":
+                    parts[4] = None
+                if parts[5] == "None":
+                    parts[5] = None
+                write_thoth_location(parts[0], parts[1], parts[2], parts[3], parts[4], parts[5])
             except IndexError:
-                raise ValueError('Not enough data in entry "{}"'.location)
+                raise ValueError('Not enough data in entry "{}"'.format(location.rstrip()))